strobelight-libs: fix segfault
There's a segfault that was happening when using the pystacks feature
with systing. We're destroying the ProcessDiscovery object and it has a
dependency on the OffsetResolver object that is also used by the
SharedPidInfoCache, and since these are destroyed in reverse order we'd
get a segfault. Fix this by swapping the order.

Signed-off-by: Josef Bacik <josefbacik@anthropic.com>