Enable 3.11 support (#2)
Summary:
We use python 3.11 in Anthropic, which has different offsets than the default (3.8). This adds the necessary offsets and changes to the BPF side to enable 3.11 support for pystacks. It also updates the default to 3.12, as anything more modern is going to be more likely to not be 3.8.

Pull Request resolved: https://github.com/facebookincubator/strobelight-libs/pull/2

Reviewed By: peremyach

Differential Revision: D88646034

Pulled By: jwiepert

fbshipit-source-id: f71c992c9aadf9a24dad3a0c5216c368ded0f038